chore(host_metrics): Break this source into modules
